+++ diagram-util/src/main/java/com/powsybl/diagram/util/forcelayout/layouts/LayoutEnum.java
@@ -6,15 +6,11 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package com.powsybl.diagram.util.forcelayout.layouts.layoutsparameters;
+package com.powsybl.diagram.util.forcelayout.layouts;
 
 /**
  * @author Nathan Dissoubray {@literal <nathan.dissoubray at rte-france.com>}
  */
-public abstract class AbstractLayoutParameters {
-    protected LayoutEnum parameterType;
-
-    public LayoutEnum getParameterType() {
-        return parameterType;
-    }
+public enum LayoutEnum {
+    SPRINGY,
 }
